pyproject.toml
[tool.poetry]
name = "chimera_ml"
version = "0.1.0"
description = "Cross-modal HIerarchical Merging of Embeddings and Representations"
authors = ["Maxim Markitantov <m.markitantov@yandex.ru>"]
readme = "README.md"
packages = [{ include = "chimera_ml", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
torch = "*"
tqdm = "*"
matplotlib = "*"
mlflow = "*"
numpy = "*"
pyyaml = "*"
pandas = "*"
scikit-learn = "*"
typer = { extras = ["all"], version = "*" }

[tool.poetry.group.dev.dependencies]
pytest = "*"
ruff = "*"

[tool.poetry.scripts]
chimera_ml = "chimera_ml.cli:app"

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"

